requirements.txt
# ============================================================
# SAPAS Python 依赖列表
# ============================================================

# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
email-validator>=2.0.0

# 数据库
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# 数据获取
akshare>=1.12.0
pandas>=2.1.0
numpy>=1.26.0

# 技术指标
pandas-ta>=0.3.14b

# 可视化
plotly>=5.18.0
kaleido>=0.2.1

# 数据校验
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 任务调度
apscheduler>=3.10.4

# 认证与安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 工具库
httpx>=0.26.0
tenacity>=8.2.3
python-dotenv>=1.0.0

# 日志
loguru>=0.7.2

# 开发依赖
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Redis (可选，用于缓存和发布订阅)
# redis>=5.0.0

